requirements.txt
## LLMs
openai


# dev
python-dotenv==1.0.0

requests